makefile
So we now have a makefile for the project. The file resides at
implementation/src/java. The default make will compile the code, create
a jar, and generate the javadocs and put the javadocs in the correct
place. make doc will generate the javadocs and put them in the right
place and make clean will remove all of the extra files from the
compilation and generating of the javadocs.